misc: add new test cases and make other changes
In addition to the new tests, also made small changes in how test
cases are presented in the README file. Now 01 or more of 03 possible
tables are shown.

- one for implemented test cases (always there)
- one for test cases to be implemented (when planning to implement)
- one for removed test cases (when such cases exist)

Before this change the second table wasn't a possibility.

Also changed the list in the "Steps" section from the test case
document template so it is an ordered list rather than an unordered
one.